Revert "Split out common compression routines into separate file (#5728)"
This reverts commit 9c88475b315541a5a86e5feb59432ab638a93d6d.
