--- stdlib/src/compression/mod.rs
@@ -1,5 +1,0 @@
-mod generic;
-use generic::*;
-
-pub mod bz2;
-pub mod zlib;
